Merge devel into master for 5.6.1 release

+++ build.sh
@@ -38,17 +38,31 @@ newer() {
   fi
 }
 
+newer_s() {
+  FILE1=$1
+  for FILE2 in "$@"
+  do
+    [ "$FILE1" = "$FILE2" ] && continue
+    if newer "$FILE2" "$FILE1"
+    then
+      return 0
+    fi
+  done
+  return 1
+}
+
 setversion() {
+  echo "Setting version to ${VERSION}"
   export LC_TIME="en_GB.UTF-8"
   
   sed -e "s/__VERSION__/${VERSION}/g" htdocs/sw.js.in > htdocs/sw.js
   sed -e "s/__VERSION__/${VERSION}/g" contrib/packaging/alpine/APKBUILD.in > contrib/packaging/alpine/APKBUILD
   sed -e "s/__VERSION__/${VERSION}/g" contrib/packaging/arch/PKGBUILD.in > contrib/packaging/arch/PKGBUILD
   DATE=$(date +"%a %b %d %Y")
-  sed -e "s/__VERSION__/${VERSION}/g" "-e s/__DATE__/$DATE/g" \
+  sed -e "s/__VERSION__/${VERSION}/g" -e "s/__DATE__/$DATE/g" \
   	contrib/packaging/rpm/mympd.spec.in > contrib/packaging/rpm/mympd.spec
   DATE=$(date +"%a, %d %b %Y %H:%m:%S %z")
-  sed -e "s/__VERSION__/${VERSION}/g" "-e s/__DATE__/$DATE/g" \
+  sed -e "s/__VERSION__/${VERSION}/g" -e "s/__DATE__/$DATE/g" \
   	contrib/packaging/debian/changelog.in > contrib/packaging/debian/changelog
 }
 
@@ -62,14 +76,16 @@ minify() {
 
   if newer "$DST" "$SRC"
   then
+    #File already minified"
     echo "Skipping $SRC"
-    return
+    return 0
   fi
   echo "Minifying $SRC"
 
   if [ "$TYPE" = "html" ] && [ "$PERLBIN" != "" ]
   then
-    $PERLBIN -pe 's/^<!--debug-->.*\n//gm; s/<!--release\s+(.+)-->/$1/g; s/^\s*//gm; s/\s*$//gm' "$SRC" | $GZIP > "$DST"
+    # shellcheck disable=SC2016
+    $PERLBIN -pe 's/^<!--debug-->.*\n//gm; s/<!--release\s+(.+)-->/$1/g; s/<!--(.+)-->//g; s/^\s*//gm; s/\s*$//gm' "$SRC" | $GZIP > "$DST"
     ERROR="$?"
   elif [ "$TYPE" = "js" ] && [ "$JAVABIN" != "" ]
   then
@@ -87,14 +103,30 @@ minify() {
   then
     echo "Error minifying $SRC, copy $SRC to $DST"
     cp "$SRC" "$DST"
+    return 2
   fi
+  #successfull minified file
+  return 1
+}
+
+createi18n() {
+  DST=$1
+  PRETTY=$2
+  cd src/i18n || exit 1
+  if newer_s "$DST" ./*.txt
+  then
+    echo "Creating i18n json"
+    $PERLBIN ./tojson.pl "$PRETTY" > "$DST"
+  else
+    echo "Skip creating i18n json"
+  fi
+  cd ../.. || exit 1
 }
 
 buildrelease() {
-  echo "Creating i18n json"
-  cd src/i18n || exit 1
-  $PERLBIN ./tojson.pl > ../../dist/htdocs/js/i18n.min.js
-  cd ../.. || exit 1
+  ASSETSCHANGED=0
+
+  createi18n ../../dist/htdocs/js/i18n.min.js
   
   echo "Minifying javascript"
   minify js htdocs/sw.js dist/htdocs/sw.min.js
@@ -103,55 +135,88 @@ buildrelease() {
   minify js htdocs/js/mympd.js dist/htdocs/js/mympd.min.js
   
   echo "Combining and compressing javascript"
-  cat dist/htdocs/js/i18n.min.js \
-      dist/htdocs/js/keymap.min.js \
-      dist/htdocs/js/bootstrap-native-v4.min.js \
-      dist/htdocs/js/mympd.min.js > dist/htdocs/js/combined.js
-  
-  $GZIPBIN -f -v -9 dist/htdocs/js/combined.js
-  $GZIPBIN -f -v -9 -c dist/htdocs/sw.min.js > dist/htdocs/sw.js.gz
+  JSFILES="dist/htdocs/js/i18n.min.js dist/htdocs/js/keymap.min.js dist/htdocs/js/bootstrap-native-v4.min.js dist/htdocs/js/mympd.min.js"
+  # shellcheck disable=SC2086
+  if newer_s dist/htdocs/js/combined.js.gz $JSFILES
+  then
+    # shellcheck disable=SC2086
+    cat $JSFILES > dist/htdocs/js/combined.js
+    $GZIPBIN -f -v -9 dist/htdocs/js/combined.js
+    ASSETSCHANGED=1
+  else
+    echo "Skip creating dist/htdocs/js/combined.js.gz"
+  fi
+  if newer dist/htdocs/sw.min.js dist/htdocs/sw.js.gz
+  then
+    $GZIPBIN -f -v -9 -c dist/htdocs/sw.min.js > dist/htdocs/sw.js.gz
+    ASSETSCHANGED=1
+  else
+    echo "Skip dist/htdocs/sw.js.gz"
+  fi
  
   echo "Minifying stylesheets"
   minify css htdocs/css/mympd.css dist/htdocs/css/mympd.min.css
   
   echo "Combining and compressing stylesheets"
-  cat dist/htdocs/css/bootstrap.css \
-      dist/htdocs/css/mympd.min.css > dist/htdocs/css/combined.css
-  
-  $GZIPBIN -f -v -9 dist/htdocs/css/combined.css
+  CSSFILES="dist/htdocs/css/bootstrap.min.css dist/htdocs/css/mympd.min.css"
+  # shellcheck disable=SC2086
+  if newer_s dist/htdocs/css/combined.css.gz $CSSFILES
+  then
+    # shellcheck disable=SC2086
+    cat $CSSFILES > dist/htdocs/css/combined.css
+    $GZIPBIN -f -v -9 dist/htdocs/css/combined.css
+    ASSETSCHANGED=1
+  else
+    echo "Skip creating dist/htdocs/css/combined.css.gz"
+  fi
   
   echo "Minifying and compressing html"
-  minify html htdocs/index.html dist/htdocs/index.html.gz
+  if minify html htdocs/index.html dist/htdocs/index.html.gz
+  then
+    ASSETSCHANGED=1
+  fi
 
   echo "Creating other compressed assets"
-  ASSETS="htdocs/mympd.webmanifest"
-  ASSETS="$ASSETS htdocs/assets/coverimage-notavailable.svg htdocs/assets/coverimage-stream.svg"
-  ASSETS="$ASSETS htdocs/assets/coverimage-loading.svg "
+  ASSETS="htdocs/mympd.webmanifest htdocs/assets/*.svg"
   for ASSET in $ASSETS
   do
     COMPRESSED="dist/${ASSET}.gz"
     if newer "$ASSET" "$COMPRESSED"
-    then 
+    then
       $GZIPBIN -v -9 -c "$ASSET" > "$COMPRESSED"
+      ASSETSCHANGED=1
+    else
+      echo "Skipping $ASSET"
     fi
   done
 
-  echo "Compiling mympd"
+  echo "Compiling myMPD"
   install -d release
   cd release || exit 1
-  #force rebuild of web_server.c with embedded assets
-  rm -f CMakeFiles/mympd.dir/src/web_server.c.o
+  if [ "$ASSETSCHANGED" = "1" ]
+  then
+    #force rebuild of web_server.c with embedded assets
+    rm -f CMakeFiles/mympd.dir/src/web_server.c.o
+  else
+    echo "Assets not changed"
+  fi
   #set INSTALL_PREFIX and build myMPD
   export INSTALL_PREFIX="${MYMPD_INSTALL_PREFIX:-/usr}"
   cmake -DCMAKE_INSTALL_PREFIX:PATH="$INSTALL_PREFIX" -DCMAKE_BUILD_TYPE=RELEASE ..
   make
 }
 
+addmympduser() {
+  echo "Checking status of mympd system user and group"
+  getent group mympd > /dev/null || groupadd -r mympd
+  getent passwd mympd > /dev/null || useradd -r -g mympd -s /bin/false -d /var/lib/mympd mympd
+}
+
 installrelease() {
-  echo "Installing mympd"
+  echo "Installing myMPD"
   cd release || exit 1  
   make install DESTDIR="$DESTDIR"
-  ../contrib/packaging/debian/postinst 
+  addmympduser
 }
 
 builddebug() {
@@ -161,12 +226,9 @@ builddebug() {
   [ -e "$PWD/htdocs/css/bootstrap.css" ] || ln -s "$PWD/dist/htdocs/css/bootstrap.css" "$PWD/htdocs/css/bootstrap.css"
   [ -e "$PWD/htdocs/js/bootstrap-native-v4.js" ] || ln -s "$PWD/dist/htdocs/js/bootstrap-native-v4.js" "$PWD/htdocs/js/bootstrap-native-v4.js"
 
-  echo "Creating i18n json"
-  cd src/i18n || exit 1
-  $PERLBIN ./tojson.pl pretty > ../../htdocs/js/i18n.js
-  cd ../.. || exit 1
+  createi18n ../../htdocs/js/i18n.js pretty
   
-  echo "Compiling"
+  echo "Compiling myMPD"
   install -d debug
   cd debug || exit 1
   cmake -DCMAKE_INSTALL_PREFIX:PATH=/usr -DCMAKE_BUILD_TYPE=DEBUG -DMEMCHECK="$MEMCHECK" ..
@@ -176,18 +238,15 @@ builddebug() {
 cleanupoldinstall() {
   if [ -d /usr/share/mympd ]
   then
-    echo "Previous installation found"
+    echo "Previous myMPD installation found"
     rm -rf /usr/share/mympd
     rm -rf /var/lib/mympd/tmp
-    mv /etc/mympd/mympd.conf /etc/mympd.conf
+    [ -f /etc/mympd/mympd.conf ] && mv /etc/mympd/mympd.conf /etc/mympd.conf
     rm -rf /etc/mympd
     rm -f /usr/lib/systemd/system/mympd.service
-    if ! ls -1qA /usr/lib/systemd/system/ | grep -q .
-    then
-      rmdir /usr/lib/systemd/system
-    fi
+    rmdir --ignore-fail-on-non-empty /usr/lib/systemd/system
   else
-    echo "No old installation found"
+    echo "No old myMPD installation found"
   fi
 }
 
@@ -208,6 +267,21 @@ cleanup() {
 
 cleanuposc() {
   rm -rf osc
+}
+
+cleanupdist() {
+  rm -f dist/htdocs/js/i18n.min.js
+  rm -f dist/htdocs/js/keymap.min.js 
+  rm -f dist/htdocs/js/bootstrap-native-v4.min.js 
+  rm -f dist/htdocs/js/mympd.min.js
+  rm -f dist/htdocs/js/combined.js.gz
+  rm -f dist/htdocs/css/mympd.min.css
+  rm -f dist/htdocs/css/combined.css.gz
+  rm -f dist/htdocs/sw.min.js
+  rm -f dist/htdocs/sw.js.gz
+  rm -f dist/htdocs/mympd.webmanifest.gz
+  rm -f dist/htdocs/index.html.gz
+  rm -f dist/htdocs/assets/*.gz
 }
 
 check () {
@@ -241,13 +315,36 @@ pkgdebian() {
   export LC_TIME="en_GB.UTF-8"
   tar -czf "../mympd_${VERSION}.orig.tar.gz" -- *
   dpkg-buildpackage -rfakeroot
+
+  #get created package name
+  PACKAGE=$(ls ../mympd_"${VERSION}"-1_*.deb)
+  if [ "$PACKAGE" = "" ]
+  then
+    echo "Can't find package"
+  fi
+
+  if [ "$SIGN" = "TRUE" ]
+  then  
+    DPKGSIG=$(command -v dpkg-sig)
+    if [ "$DPKGSIG" != "" ]
+    then
+      if [ "$GPGKEYID" != "" ]
+      then
+        echo "Signing package with key $GPGKEYID"
+        dpkg-sig -k "$GPGKEYID" --sign builder "$PACKAGE"
+      else
+        echo "WARNING: GPGKEYID not set, can't sign package"
+      fi
+    else
+      echo "WARNING: dpkg-sig not found, can't sign package"
+    fi
+  fi
+  
   LINTIAN=$(command -v lintian)
   if [ "$LINTIAN" != "" ]
   then
     echo "Checking package with lintian"
-    ARCH=$(uname -p)
-    [ "$ARCH" = "x86_64" ] && ARCH="amd64"
-    $LINTIAN "../mympd_${VERSION}-1_${ARCH}.deb"
+    $LINTIAN "$PACKAGE"
   else
     echo "WARNING: lintian not found, can't check package"
   fi
@@ -296,9 +393,14 @@ pkgrpm() {
 pkgarch() {
   prepare
   tar -czf "mympd_${VERSION}.orig.tar.gz" -- *
-  [ "$1" = "taronly" ] && return 0
   cp contrib/packaging/arch/* .
   makepkg
+  if [ "$SIGN" = "TRUE" ]
+  then
+    KEYARG=""
+    [ "$GPGKEYID" != "" ] && KEYARG="--key $PGPGKEYID"
+    makepkg --sign "$KEYARG" mympd-*.pkg.tar.xz
+  fi
   NAMCAP=$(command -v namcap)
   if [ "$NAMCAP" != "" ]
   then
@@ -311,6 +413,13 @@ pkgarch() {
 }
 
 pkgosc() {
+  OSCBIN=$(command -v osc)
+  if [ "$OSCBIN" = "" ]
+  then
+    echo "ERROR: osc not found"
+    exit 1
+  fi
+  
   cleanup
   cleanuposc
   [ "$OSC_REPO" = "" ] && OSC_REPO="home:jcorporation/myMPD"
@@ -335,19 +444,20 @@ pkgosc() {
   fi
 
   cd "$STARTPATH/osc" || exit 1
-  cp ../package/mympd_${VERSION}.orig.tar.gz "$OSC_REPO/"
+  cp "../package/mympd_${VERSION}.orig.tar.gz" "$OSC_REPO/"
   if [ -f /etc/debian_version ]
   then
-    cp ../package/mympd_${VERSION}-1.dsc "$OSC_REPO/"
-    cp ../package/mympd_${VERSION}-1.debian.tar.xz  "$OSC_REPO/"
+    cp "../package/mympd_${VERSION}-1.dsc" "$OSC_REPO/"
+    cp "../package/mympd_${VERSION}-1.debian.tar.xz"  "$OSC_REPO/"
   fi
   cp ../contrib/packaging/rpm/mympd.spec "$OSC_REPO/"
   cp ../contrib/packaging/arch/PKGBUILD "$OSC_REPO/"
   cp ../contrib/packaging/arch/archlinux.install "$OSC_REPO/"
 
-  cd "$OSC_REPO"
+  cd "$OSC_REPO" || exit 1
   osc addremove
   osc st
+  osc vc
   osc commit
 }
 
@@ -378,6 +488,9 @@ case "$1" in
 	  cleanup
 	  cleanuposc
 	;;
+	cleanupdist)
+	  cleanupdist
+	;;
 	check)
 	  check
 	;;
@@ -402,33 +515,51 @@ case "$1" in
 	pkgosc)
 	  pkgosc
 	;;
+	addmympduser)
+	  addmympduser
+	;;
 	*)
 	  echo "Usage: $0 <option>"
 	  echo "Version: ${VERSION}"
-	  echo "Options:"
+	  echo ""
+	  echo "Build options:"
 	  echo "  release:        build release files in directory release"
 	  echo "                  following environment variables are respected"
-	  echo "                  MYMPD_INSTALL_PREFIX=/usr"
+	  echo "                    - MYMPD_INSTALL_PREFIX=\"/usr\""
 	  echo "  install:        installs release files from directory release"
 	  echo "                  following environment variables are respected"
-	  echo "                  DESTDIR=\"\""
+	  echo "                    - DESTDIR=\"\""
 	  echo "  releaseinstall: calls release and install afterwards"
 	  echo "  debug:          builds debug files in directory debug"
 	  echo "                  linked with libasan3, uses assets in htdocs"
 	  echo "  memcheck:       builds debug files in directory debug"
 	  echo "                  for use with valgrind, uses assets in htdocs/"
-	  echo "  cleanupoldinst: removes deprecated files"
+	  echo "  check:          runs cppcheck on source files"
+	  echo ""
+	  echo "Cleanup options:"
 	  echo "  cleanup:        cleanup source tree"
-	  echo "  check:          runs cppcheck"
+	  echo "  cleanupdist:    cleanup dist directory, forces release to build new assets"
+	  echo "  cleanupoldinst: removes deprecated files"
+	  echo ""
+	  echo "Packaging options:"
 	  echo "  pkgalpine:      creates the alpine package"
 	  echo "  pkgarch:        creates the arch package"
+	  echo "                  following environment variables are respected"
+	  echo "                    - SIGN=\"FALSE\""
+	  echo "                    - GPGKEYID=\"\""
 	  echo "  pkgdebian:      creates the debian package"
+	  echo "                  following environment variables are respected"
+	  echo "                    - SIGN=\"FALSE\""
+	  echo "                    - GPGKEYID=\"\""
 	  echo "  pkgdocker:      creates the docker image (debian based with libmpdclient from git master branch)"
 	  echo "  pkgrpm:         creates the rpm package"
+	  echo "  pkgosc:         updates the open build service repository"
+	  echo "                  following environment variables are respected"
+	  echo "                    - OSC_REPO=\"home:jcorporation/myMPD\""
+	  echo ""
+	  echo "Misc options:"
 	  echo "  setversion:     sets version and date in packaging files from CMakeLists.txt"
-#Working on osc support
-#	  echo "  pkgosc:            updates a osc repository"
-#	  echo "                  following environment variables are respected"
-#	  echo "                  OSC_REPO=\"home:jcorporation/myMPD\""
+	  echo "  addmympduser:   adds mympd group and user"
+	  echo ""
 	;;
 esac
